requirements.txt
customtkinter==5.1.2
dropbox==11.36.2
pyzipper==0.3.6
PyDrive2==1.18.0
humanize==4.6.0
psutil==6.0.0
Pillow==11.0.0
plyer==2.1.0
schedule==1.2.2
prettytable==3.7.0
colorama==0.4.6
art==5.9
Flask==2.3.3
Werkzeug==2.3.7
